requirements.txt
absl-py==1.4.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.1.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.5
distro==1.4.0
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
expecttest==0.1.3
fabric==2.7.1
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.34.4
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==0.4.15
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.1
identify==2.5.1
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.8.0
intel-openmp==2022.1.0
invoke==1.7.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2022.1.0
mkl-devel==2022.1.0
mkl-include==2022.1.0
mlnx-tools===-5.2.0-
mpi4py==3.1.3
msgpack==0.6.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
neovim==0.3.1
nest-asyncio==1.5.5
nodeenv==1.7.0
notebook==6.4.12
numpy==1.23.0
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.22.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==2.0.3
PyNaCl==1.5.0
pynvim==0.4.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.10.31
requests==2.28.0
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.14.0
smmap==5.0.0
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.3.0
tbb==2021.6.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
tinycss2==1.1.1
titans==0.0.5
tokenizers==0.13.2
toml==0.10.2
torchvision==0.14.0a0+9effc4c
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.26.1
types-dataclasses==0.6.5
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.15.1
wandb==0.13.11
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.6.1
zipp==3.8.0
